Implement email-password auth,  middleware, and start on layout (navbar)
The email-password authentication is fully working but has not been tested or iterated thoroughly yet, so definitely needs a revisit. Also the login-form is functional but not styled yet.

Middleware has been implemented so that the application is not accessible unless the user has logged in.

The layout (structure and styling) has been started, a first version of the NavBar has been made and styled. Right now it has static navbar links, but eventually this should be dynamic based on configs.