Establish folder structure, first start on Teams page
Started filling out the app folder with the different routes and populate their folders with a relevant folders like components, services, etc.

Also started on the Teams Page: Teams Page now has a TeamsList, and a TeamsList includes multiple TeamListItems. Upon clicking a TeamListItem, the TeamDetails page comes sliding in. This is a great start on the structure and functionality of this feature, but the content is not yet finalized and will have to be looked at later.