Start made on the admin part of dashboard
The folder structure of the admin pages has been prepared and navbar has been altered to be able to show different links when on an admin page than when not, and also the background color has changed. It's not done in the best way yet, it can be improved by making lists/arrays of the links for admin vs normal pages, and the classnames can be done using clsx.