Big changes to projects structure
The project now has two main routes it can take: auth and main. When not logged in yet, the middleware redirects the user to the (auth)/login route. This route does not have a navbar, only a login form. When logged in, the middleware sends the user to the route '/'. This is located at (main)/page.tsx.

the app folder now also only includes route pages, and their styling files.