Create API for committee data on MemberDetailPage
The current committees and committees history for a member are now fetched and displayed on their own designated locations.

As of now, all of the member detail page information (team and committee info) are fetched through a Promise.all, meaning waiting times are semi-long. In the future this might have to be revisited.